Add MusicSession animations

+++ src/include/Components/UI/Image.h
@@ -26,7 +26,7 @@ public:
     * y from 0 to 1080
     */
     void LoadTexture(int x, int y, const std::string& path);
-    void SetPosition(int x2, int y2);
+    void SetPosition(float x2, float y2);
     void SetRotation(float angle);
     void SetScale(float scale);
